Smart-anomly-detector/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
shap
shapely
geopandas
mlflow
sqlalchemy
python-multipart
python-dateutil
psycopg2-binary
python-dotenv